+++ server/BusinessLogic/FieldsHandler.js
@@ -1,5 +1,7 @@
 //Function used to check the request fields (valid emails, valid field lengths and valid text fields)
 var validator = require('validator');
+// Import logger
+const logger = require('./../utils/logger/logger');
 async function check_fields(req){
     data = req.body.request
     for (const key of Object.keys(data)) {
